requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
crispy-bootstrap5==0.7
cryptography==41.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-allauth==0.54.0
django-bootstrap-icons==0.8.3
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-filter==23.2
django-import-export==3.2.0
django-phonenumber-field==7.1.0
et-xmlfile==1.1.0
icalendar==5.0.7
idna==3.4
MarkupPy==1.14
mysqlclient==2.2.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
phonenumberslite==8.13.19
Pillow==10.0.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==5.5.0
tablib==3.5.0
typing_extensions==4.7.1
urllib3==2.0.4
whitenoise==6.5.0
xlrd==2.0.1
xlwt==1.3.0
